Cora Dataset
CORA ML
The Cora dataset consists of Machine Learning papers.
 These papers are classified into one of the following seven classes:
  Case_Based
  Genetic_Algorithms 
 Neural_Networks 
 Probabilistic_Methods
 Reinforcement_Learning 
 Rule_Learning 
 Theory

The file contains the citation graph of the corpus. 
Each line describes a link in the following format:       Each line contains two paper IDs. 

         The first entry is the ID of the paper being cited 
        The second ID stands for the paper which contains the citation. 

The direction of the link is from right to left. If a line is represented by "paper1 paper2" then the link is "paper2->paper1